GremlinServer uses ConfigurationGraphManagement
This is because this commit includes a graph with a key
"JanusConfigurationGraph". The graph is defined by the supplied
properties file, and in this example, the graph hits a local
cassandrathrift on a keyspace "JanusConfigurationGraph", as configured
because that is what the graph.graphname is set to and no keyspace is
supplied.