UpdateTriggerStatus from Set -> List
Previously, we were seeing new index creations on the same graph act as
if they were equal to previous triggers. Thus, the previous but
not finished updateStatusTriggers were being removed from the SET and
getting stuck in INSTALLED. In this change, we have done two things:
1. Change the datastructure housing the updateStatusTriggers from a Set
to a List.
2. Remove the code that removes any updateStatusTriggers from the data
structure.

These changes ensure all index creations on a given graph fulfill their
INSTALLing.