Create JanusGraphManager
This class implements Tinkerpop's GraphManager, and allows for reference
tracking of a Graph references stored in a ConcurrentHashMap<String,
Graph>.

This class is instantiated by Tinkerpop's GremlinServer and acts as a
Singleton, hence the RuntimeException that gets thrown upon a second
instantiation attempt.

The openGraph() and removeGraph() methods allow for thread-safe
coordination to instantiate graph references, add them to the reference
tracker, and remove them from the reference tracker.

Signed-off-by: David Pitera <dpitera@us.ibm.com>