Paginate column access in AstyanaxStore
JanusGraph has a tendency to make queries to Cassandra with _very_ large
limits. For example, any query without a `.limit()` clause will fetch
all matching columns from a given row. These requests are usually
range-bound (e.g., only requesting columns that are edges, based on the
serialization logic) but are not limited.

Fetching millions of columns in a single request is not very friendly to
Cassandra; in production we've observed memory allocation spikes in
excess of 10GB in a 30s period. On my local machine, calling

graph.traversal().V(id).outE().count()

prior to this commit will cause my Cassandra instance to OOM in
approximately 10 seconds.

This patch places a hard cap on the number of columns returned (per row)
via a single request to Cassandra. If, for a given request, a row has
more columns available in the requested range than the cap, the system
makes additional requests for that row until the requested number of
columns is returned or no more columns are available.

This code is somewhat more subtle that may be expected due to JanusGraph's
use of the Astyanax API - it's not uncommon for JanusGraph to make a request
for multiple keys concurrently, in which case the `limit` parameter
applies to each key - not the request as a whole. This has two
implications:

1.  We need to use a different API (getRow, rather than getKeySlice) to
fetch columns for rows that met the limit on the initial query, since
the starting column must be manipulated on a per-row basis

2. The page size needs to be relatively conservative, in case a single
request is made for hundreds of keys at once. This is relatively
unlikely, especially given the natural infrequency of supernodes,
but it's something that we must be aware of.

Signed-off-by: David Pitera <dpitera@us.ibm.com>